Normalize all registry IDs to uint32 for AutoIDRegistry prep
Mechanical refactor: convert 17 int-keyed registries in ObjectRegistry
to uint32 keys with atomic.Uint32 counters. All converter functions
(xxxToValue/xxxIDFromValue) updated to match, removing unnecessary
int↔uint32 casts throughout.

Also fixes latent bug in RegisterException: Add(1) without -1 meant
first exception ID was 2 instead of 1.

No behavior change (except exception ID fix). Prepares for commit 2
which introduces AutoIDRegistry[V] to collapse the per-domain
boilerplate.

+++ vm/class_value.go
@@ -12,8 +12,8 @@ package vm
 // *Class pointer in the VM-local class value registry (ObjectRegistry).
 
 // classToValue creates a Value from a class registry ID.
-func classToValue(id int) Value {
-	return FromSymbolID(uint32(id) | classValueMarker)
+func classToValue(id uint32) Value {
+	return FromSymbolID(id | classValueMarker)
 }
 
 // isClassValue returns true if this value is a class value.
@@ -26,8 +26,8 @@ func isClassValue(v Value) bool {
 }
 
 // classValueIDFromValue extracts the registry ID from a class value.
-func classValueIDFromValue(v Value) int {
-	return int(v.SymbolID() & ^markerMask)
+func classValueIDFromValue(v Value) uint32 {
+	return v.SymbolID() & ^markerMask
 }
 
 // ---------------------------------------------------------------------------
